WIP: ENGINE-524: make paths to sys, mgmt and keys DB configurable via env
Rationale: full compatibility on systems which already have a system and management DB,
while moving things under $HOME/.pEp (or $XDG_CONFIG_DIR/pEp) on new sytems.

TODO:

 - Write and run tests
 - Do we need to create sub-directories (e.g. ~/.pEp)?
 - Bring the new defines into the Makefiles
 - Implementation for Windows: windoze_local_db()

Changes:

 - The hardcoded location for system DB is now in /usr/local/share/pEp instead of
   /usr/share/pEp, the Makefile should default to /usr/share/pEp.


User documentation:

Set the following environment variables to select the location of the various databases:

TRUSTWORDS sets path to:
 - $TRUSTWORDS/system.db

PEPHOME sets path to:
 - $PEPHOME/pEp_management.db (*)
 - $PEPHOME/pEp_keys.db (**)
 - $PEPHOME/system.db (***)

(*) Note that there is no dot in the filename (pEp_management.db) when PEPHOME
is set. The default is still $HOME/.pEp_management.db (with dot) when PEPHOME
is unset.

(**) When you set PEPHOME, both pEp_keys.db and pEp_management.db must be
co-located. If the databases don't exist, new ones will be initialized. The
PEPHOME directory must exist before starting the adapters.

(***) system.db will be found in PEPHOME only if TRUSTWORDS is not set. If
TRUSTWORDS is set, the system DB MUST be where TRUSTWORDS points to. Otherwise,
if TRUSTWORDS is unset and PEPHOME contains no system DB, search will continue
at the default location. In other words, while keys and management DB must be
in PEPHOME when set, the system DB may be in the default location too.

The default locations are:
 - $HOME/.pEp/pEp_management.db (if file exists)
 - $HOME/.pEp_management.db (otherwise)
 - /usr/local/share/pEp/system.db
 - $HOME/.pEp_keys.db