CI: add basic gitlab-ci jobs
Adds simple jobs for building the the Engine for Debian 10.

Pre-built docker images containing Sequoia will be pulled down and used
to build the Engine. Currently the Sequoia version is just the latest
from the `pep-engine` branch. We should make this selectable in the near
future.

Upon successful compilation an image containing the build output will be
pushed to an internal docker registry.