message struct: add rating field
Add the new rating field to the ASN.1 representation as well (I am about to
discover, in a forthcoming commit, that that is in fact a stupid decision).

Move the rating enum definition from message_api.h to pEpEngine.h in order to
make it accessible from message.h .  This is temporary anyway: the new rating
engine is coming.

Remove the rating output parameter from decrypt_message.  This will temporarily
break *everything*!