media key: fix bug in config_media_keys reported by dirk
config_media_keys: Before adding the new entries to the map, make the map null.
The code is slightly complicated by the need to keep the *old* map alive in
order to handle the out-of-memory case cleanly, but this was still a very ugly
bug.  Thanks Dirk for reporting it.

Thanks dirk for reporting.