BUILD: Fix - Effectively add include and lib dir from local.conf
the -I and -L flags arent set correctly currently
basically the local.conf file PREFIX= gets ignored
how it should be
-I$(PREFIX)/include
-L$(PREFIX)/lib
but this NEVER was the case.
it just worked because of a side effect when using sequoia as the crypto backend.
there we do e.g: SEQUOIA_LDFLAGS+=$(shell pkg-config --libs-only-L pep_engine_sequoia_backend)