PEMA-103 Fixed and improved android build
Updated .gitignore (no longer takes android include folder, which holds 'autogen' data).

Updated platform_unix.c, android requires uuid.h to be included inside of platform_unix.c and does not use uuid_t[16] as MVS does this was a typo. MVS did not have this filled, and since I do not develop for MVS I have left it empty on purpose for MVS. 
Note, this should fix the FIXME issue, I have not removed it since I am not 100% sure what timegm means in this context. 

Note 2: I know the if nest there is a bit of a code smell, to put it mildly, but I can't think of a better way to do it, without accidentally breaking MVS.

Updated platform_unix.h once more as I am not privvy (yet) to the logic inside of MVS builds, I have removed the android check to add unsigned char uuid_t[16] and added !defined check on the include for uuid/uuid.h. To avoid adding uuid/uuid.h on the .h file, which breaks android, android adds nothing and then goes into the if-else logic.

Updated SQLite3 amalgam, it was not updated since a long time ago, as most developers seem to use the system standard, on android however because of NDK-adjacent issues we must generate our own, which means this must be updated. Current update is v3.39 in lieu of an exact version.