CORE-240: Re-Propagate updated public key
The most significant change in this patch is that PEP_ct_mistrusted
is now being upgraded and handled like PEP_ct_key_expired in validate_fpr.
This change is nessecary so that a prior but reset identity can be resumed
from the database in order to communicate the update of the key using which
needs to be signed with the old key for security reasons.