CORE-240: Re-Propagate updated public key
The most significant change in this patch is the introduction of a
new flag, called PEP_encrypt_flag_key_reset_repropagate which is meant
to signify the engine, that a previously revoked key might be going to
be needed in order to notify partners of the updated identity.

This however will obviously not work, when the other side has already
received a revocation certificate and doesn't accept your key anymore,
which is a case which will never happen, because if they've revoked your
key, they certainly won't encrypt and email using it.